refactor(Auth): 인증 코드 받는거부터 시작하기 위한 세팅
