--- src/main/java/ddd/caffeine/ratrip/module/auth/application/AuthService.java
@@ -1,11 +1,0 @@
-package ddd.caffeine.ratrip.module.auth.application;
-
-import ddd.caffeine.ratrip.module.auth.application.dto.SignInDto;
-import ddd.caffeine.ratrip.module.auth.application.dto.SignUpDto;
-import ddd.caffeine.ratrip.module.auth.presentation.dto.response.SignInResponseDto;
-
-public interface AuthService {
-	SignInResponseDto signUp(SignUpDto request);
-
-	SignInResponseDto signIn(SignInDto request);
-}
