--- src/main/java/ddd/caffeine/ratrip/module/auth/application/AuthServiceProvider.java
@@ -1,29 +1,0 @@
-package ddd.caffeine.ratrip.module.auth.application;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import javax.annotation.PostConstruct;
-
-import org.springframework.stereotype.Component;
-
-import ddd.caffeine.ratrip.module.auth.application.impl.KakaoAuthService;
-import ddd.caffeine.ratrip.module.user.domain.UserSocialType;
-import lombok.RequiredArgsConstructor;
-
-@RequiredArgsConstructor
-@Component
-public class AuthServiceProvider {
-	private static final Map<UserSocialType, AuthService> authServiceMap = new HashMap<>();
-
-	private final KakaoAuthService kakaoAuthService;
-
-	@PostConstruct
-	void initializeAuthServicesMap() {
-		authServiceMap.put(UserSocialType.KAKAO, kakaoAuthService);
-	}
-
-	public AuthService getAuthService(UserSocialType socialType) {
-		return authServiceMap.get(socialType);
-	}
-}
