+++ src/main/java/ddd/caffeine/ratrip/module/auth/presentation/AuthController.java
@@ -10,12 +10,8 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import ddd.caffeine.ratrip.module.auth.application.AuthService;
-import ddd.caffeine.ratrip.module.auth.application.AuthServiceProvider;
 import ddd.caffeine.ratrip.module.auth.application.TokenService;
-import ddd.caffeine.ratrip.module.auth.presentation.dto.request.SignInRequestDto;
 import ddd.caffeine.ratrip.module.auth.presentation.dto.request.SignOutRequestDto;
-import ddd.caffeine.ratrip.module.auth.presentation.dto.request.SignUpRequestDto;
 import ddd.caffeine.ratrip.module.auth.presentation.dto.request.TokenReissueRequestDto;
 import ddd.caffeine.ratrip.module.auth.presentation.dto.response.SignInResponseDto;
 import ddd.caffeine.ratrip.module.auth.presentation.dto.response.TokenResponseDto;
@@ -25,19 +21,11 @@ import lombok.RequiredArgsConstructor;
 @RequestMapping("/v1")
 @RequiredArgsConstructor
 public class AuthController {
-	private final AuthServiceProvider authServiceProvider;
 	private final TokenService tokenService;
 
-	@PostMapping("/auth/signup")
-	public ResponseEntity<SignInResponseDto> signUp(@Valid @RequestBody SignUpRequestDto request) {
-		AuthService authService = authServiceProvider.getAuthService(request.getSocialType());
-		return ResponseEntity.ok(authService.signUp(request.toServiceDto()));
-	}
-
-	@PostMapping("/auth/signin")
-	public ResponseEntity<SignInResponseDto> signIn(@Valid @RequestBody SignInRequestDto request) {
-		AuthService authService = authServiceProvider.getAuthService(request.getSocialType());
-		return ResponseEntity.ok(authService.signIn(request.toServiceDto()));
+	@PostMapping("/auth/signin/kakao")
+	public ResponseEntity<SignInResponseDto> signInWithKakao() {
+		return ResponseEntity.ok();
 	}
 
 	@PostMapping("/auth/reissue")
